Cleaned up PlayerMoveEvent and PlayerTeleportEvent.  Now both inherit from a new PlayerChangePositionEvent.
